+++ indico/web/client/js/react/components/notes/NoteEditor.jsx
@@ -5,45 +5,101 @@
 // modify it under the terms of the MIT License; see the
 // LICENSE file for more details.
 
+import renderMarkdownURL from 'indico-url:core.markdown';
+import userPreferencesMarkdownAPI from 'indico-url:users.user_preferences_markdown_api';
+
 import _ from 'lodash';
 import PropTypes from 'prop-types';
-import React, {useState} from 'react';
-import {Dimmer, Loader} from 'semantic-ui-react';
+import React, {useState, useEffect, useCallback} from 'react';
+import {Field} from 'react-final-form';
+import {Button, Dimmer, Loader} from 'semantic-ui-react';
 
 import {handleSubmitError} from 'indico/react/forms';
 import {FinalModalForm} from 'indico/react/forms/final-form';
 import {Translate} from 'indico/react/i18n';
 import {indicoAxios, handleAxiosError} from 'indico/utils/axios';
+import {camelizeKeys} from 'indico/utils/case';
 
+import {FinalMarkdownEditor} from '../MarkdownEditor';
 import {FinalTextEditor} from '../TextEditor';
+
+// import style manually
+import 'react-markdown-editor-lite/lib/index.css';
 
 export function NoteEditor({apiURL, imageUploadURL, closeModal, getNoteURL}) {
   const [currentInput, setCurrentInput] = useState(undefined);
   const [loading, setLoading] = useState(false);
   const [allowWithoutChange, setAllowWithoutChange] = useState(false);
+  const [renderMode, setRenderMode] = useState(undefined);
 
-  const getNote = async () => {
+  const combineNotes = (resp, mode) => {
+    if (mode === 'markdown') {
+      setCurrentInput(
+        resp.data.notes
+          .map(note => `# ● ${_.escape(note.objectTitle)}\n${note.source}`)
+          .join('\n\n---\n\n')
+      );
+    } else {
+      setCurrentInput(
+        resp.data.notes
+          .map(note => `<h1>● ${_.escape(note.objectTitle)}</h1>\n${note.html}`)
+          .join('\n<hr>\n')
+      );
+    }
+    setAllowWithoutChange(true);
+  };
+
+  const getUserPreferences = async () => {
+    let resp;
+    try {
+      resp = await indicoAxios.get(userPreferencesMarkdownAPI());
+    } catch (error) {
+      handleAxiosError(error);
+      return;
+    }
+    const mode = resp.data ? 'markdown' : 'html';
+    setRenderMode(mode);
+    setLoading(false);
+    return mode;
+  };
+
+  const getNote = useCallback(async () => {
     let resp;
     try {
       resp = await indicoAxios.get(getNoteURL || apiURL);
+      resp = camelizeKeys(resp);
     } catch (error) {
       handleAxiosError(error);
       return;
     }
     if (resp.data.source) {
+      // in case we only edit minutes:
       setCurrentInput(resp.data.source);
+      if (resp.data.renderMode) {
+        // in this case we should always get the render mode anyway, so this might not be needed
+        setRenderMode(resp.data.renderMode);
+        setLoading(false);
+      } else {
+        getUserPreferences();
+      }
     } else if (resp.data.notes) {
-      setCurrentInput(
-        resp.data.notes
-          .map(note => `<h1>● ${_.escape(note.object_title)}</h1>\n${note.html}`)
-          .join('\n<hr>\n')
-      );
-      setAllowWithoutChange(true);
+      // in case we compile minutes:
+      if (resp.data.notes.every(n => n.renderMode === 'markdown')) {
+        // we will use markdown for compiling exactly for one case:
+        // all minutes are markdown and the user has a preference for writing minutes in markdown
+        const mode = await getUserPreferences();
+        combineNotes(resp, mode);
+      } else {
+        // in all other cases we want to use html
+        combineNotes(resp, 'html');
+      }
+      setLoading(false);
     } else {
+      // in case we create new minutes:
       setCurrentInput('');
+      getUserPreferences();
     }
-    setLoading(false);
-  };
+  }, [apiURL, getNoteURL]);
 
   const handleSubmit = async data => {
     try {
@@ -52,7 +108,7 @@ export function NoteEditor({apiURL, imageUploadURL, closeModal, getNoteURL}) {
       } else {
         await indicoAxios.post(apiURL, {
           source: data.source,
-          render_mode: 'html',
+          render_mode: renderMode,
         });
       }
     } catch (e) {
@@ -64,10 +120,22 @@ export function NoteEditor({apiURL, imageUploadURL, closeModal, getNoteURL}) {
     await new Promise(() => {});
   };
 
-  if (currentInput === undefined && !loading) {
+  const convertToHTML = markdownSource => async () => {
+    let resp;
+    try {
+      resp = await indicoAxios.post(renderMarkdownURL(), {source: markdownSource});
+    } catch (e) {
+      handleAxiosError(e);
+      return;
+    }
+    setRenderMode('html');
+    setCurrentInput(resp.data.html);
+  };
+
+  useEffect(() => {
     setLoading(true);
     getNote();
-  }
+  }, [getNote]);
 
   return (
     <>
@@ -85,14 +153,34 @@ export function NoteEditor({apiURL, imageUploadURL, closeModal, getNoteURL}) {
           onSubmit={handleSubmit}
           disabledUntilChange={!allowWithoutChange}
           unloadPrompt
+          extraActions={fprops =>
+            renderMode === 'markdown' && (
+              <Field name="source" subscription={{value: true}}>
+                {({input: {value: markdownSource}}) => (
+                  <Button
+                    style={{marginRight: 'auto'}}
+                    onClick={convertToHTML(markdownSource)}
+                    disabled={fprops.submitting}
+                    basic
+                  >
+                    <Translate>Change editor to HTML</Translate>
+                  </Button>
+                )}
+              </Field>
+            )
+          }
         >
-          <FinalTextEditor
-            name="source"
-            loading={loading}
-            value={currentInput}
-            parse={v => v}
-            config={{images: true, imageUploadURL}}
-          />
+          {renderMode === 'markdown' ? (
+            <FinalMarkdownEditor name="source" imageUploadURL={imageUploadURL} />
+          ) : (
+            <FinalTextEditor
+              name="source"
+              loading={loading}
+              value={currentInput}
+              parse={v => v}
+              config={{images: true, imageUploadURL}}
+            />
+          )}
         </FinalModalForm>
       )}
     </>