+++ CHANGES.rst
@@ -14,6 +14,7 @@ Improvements
 - Restore the "fullscreen view" option in CKEditor (:pr:`5505`)
 - Display & enforce judging deadline (:pr:`5506`)
 - Add a setting to disable entering persons in person link fields manually (:pr:`5499`)
+- Allow taking minutes in markdown (:issue:`3386`, :pr:`5500`, thanks :user:`Leats`)
 
 Bugfixes
 ^^^^^^^^